+++ core/perlinNoise/PerlinNoise.h
@@ -18,7 +18,7 @@ namespace WorldMaker {
         double dotGradient(int x, int y, double sampleX, double sampleY);
     public:
         PerlinNoise(int width, int height, double frequency, uint64_t seed);
-        double getPerlinNoise(int x, int y);
+        double getPerlinNoise(float x, float y);
     };
 
     class PerlinNoise3D {
@@ -31,7 +31,7 @@ namespace WorldMaker {
         double dotGradient3D(int x, int y, int z, double sampleX, double sampleY, double sampleZ);
     public:
         PerlinNoise3D(int width, int height, double frequency, double heightScale, uint64_t seed);
-        double getPerlinNoise3D(int x, int y, int z);
+        double getPerlinNoise3D(float x, float y, float z);
     };
 
     class FractalNoise {
@@ -41,7 +41,7 @@ namespace WorldMaker {
 
     public:
         FractalNoise(int width, int height, double frequency, double amplitude, uint64_t seed, int octaves, double lacunarity, double persistence);
-        double getNoise(int x, int y);
+        double getNoise(float x, float y);
     };
 
 
@@ -51,7 +51,7 @@ namespace WorldMaker {
         std::vector<double> m_amplitudes = {};
     public:
         ComplexNoise(int width, int height, int first, const std::vector<double>& amplitudes, uint64_t seed);
-        double getNoise(int x, int y);
+        double getNoise(float x, float y);
     };
 
 }
